+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## In-development
 - Modified the in-game trades to be more useful
-- Added Swarm mechanic
+- Implemented the Swarm mechanic (Including swarm Pokemon showing up in the Pokedex correctly)
 - Increased the variety of wild Pokemon in some areas
 
 ## 0.1.0